requirements.txt
numpy>=1.11.0
scipy>=0.17.0
nose>=1.3.7
Pillow>=3.1.1
Theano>=0.8.1
git+git://github.com/rakeshvar/theanet.git
